Changement arbo + ajout des notes
