+++ MyGes.py
@@ -1,50 +1,59 @@
-import requests #Enable HTTP requests
-from requests import*
-from bs4 import BeautifulSoup #Find Elements on website 
-import mechanize #Permet de remplir les formulaires automatiquement
-def afficher_notes():
-    #Partie connexion
-    url = "https://myges.fr/student/marks;JSESSIONID=2F17CB8504EC86A839C644842DFE57A5" #URL du site à parcourir
-    user = ""
-    mdp = ""
-    page = requests.get(url, auth=('user', 'mdp'))
-    code_connexion = str(page.status_code)
-    if code_connexion == "200":
-        print("Connexion success")
-        print("Status code: ", page.status_code)
-    else:
-        print("Connexion failed")
-        print(f"ERROR: {code_connexion}")
-        exit()
+#!/usr/bin/env python3
+
+import requests
+from urllib.parse import urlsplit, parse_qs
+import base64
+import json
+import time
 
 
-    #Partie scrapping
-    soup = BeautifulSoup(page.content, 'html.parser')
-    print(soup.title) #Permet d'afficher toute la page
-    print("Récupération de vos notes en cours...")
-    matrice_notes = []
-    #Récupération des titres
-    liste_titres = soup.find_all("span", class_="ui-column-title")
-    div_notes = soup.find_all("tr", role="row")
-    for i in liste_titres:
-        if i == None:
-            print("La case est vide")
-        else:
-            print(i.string, end=" ")
+class MYGES:
+    def __init__(self, username, password):
+        self.username = username
+        self.password = password
+        self.actionurl = "https://api.kordis.fr"
+        self.token = self.get_token()
 
-    print("\nSTOP TITRE\n")
-    #Recuperation des matieres
-    notes = soup.find_all("tr", class_="ui-widget-content")
-    for i in notes:
+    def get_token(self):
+        urltoken = 'https://authentication.kordis.fr/oauth/authorize?response_type=token&client_id=skolae-app'
+        token = base64.b64encode(f"{self.username}:{self.password}".encode()).decode('UTF-8')
+        headers = {"Authorization": f"Basic {token}"}
+        try:
+            r = requests.get(urltoken, headers=headers)
+        except requests.exceptions.InvalidSchema as e:
+            urltokenreturn = repr(e).split("'")[1].replace("#", '?')
+
+        tokenret = {k: v[0] for k, v in parse_qs(urlsplit(urltokenreturn).query).items()}
+        return {"Authorization" : f"{tokenret['token_type']} {tokenret['access_token']}"}
+    
+    def get_absences(self, year):
+        return requests.get(f"{self.actionurl}/me/{year}/absences", headers=self.token).json()
         
-        if i == None:
-            print("La case est vide")
-        else:
-            print(i.td.string)
 
-def mist(): #Affiche les ABSENCES ET les RETARDS
-    print("test")
-afficher_notes()
+    def print_absences(self, year):
+        jsondata = self.get_absences(year)
+        for row in jsondata["result"]:
+            date = time.strftime('%d-%m-%Y %H:%M', time.localtime(int(str(row['date'])[:-3])))
+            just = "Jusitifiée" if row["justified"] else "Non justifiée"
+            print(f'{date}\t{just}\t{row["course_name"]}')
+    
+    def get_agenda(self, start, end):
+        return requests.get(f"{self.actionurl}/me/agenda?start={start}&end={end}", headers=self.token)
 
-#Notes <tr data-ri="0" class="ui-widget-content ui-datatable-even odd-row" role="row">
-#Abscences s<tr data-ri="1" class="ui-widget-content ui-datatable-odd even-row" role="row">
+    def get_grades(self, year):
+        return requests.get(f"{self.actionurl}/me/{year}/grades", headers=self.token).json()
+    
+    def print_grades(self, year):
+         jsondata = self.get_grades(year)
+         for row in jsondata["result"]:
+            nom_cours = row["course"]
+            nom_prof = row["teacher_last_name"]
+            grades = row["grades"] if row["grades"] else "Pas de note"
+            print(f'{nom_cours}\t{nom_prof}\t{grades}')
+
+def main():
+    myges = MYGES("rmassiet", "ta3JHeK8")
+    myges.print_absences("2021")
+    myges.print_grades("2021")
+if __name__ == '__main__':
+    main()
